+++ Swarm/FollowPath.cs
@@ -53,7 +53,7 @@ namespace MissionPlanner.Swarm
                         alt = (float) target.Alt,
                         lat = target.Lat,
                         lng = target.Lng,
-                        id = (byte) MAVLink.MAV_CMD.WAYPOINT
+                        id = (ushort)MAVLink.MAV_CMD.WAYPOINT
                     });
                 }
             }
